+++ CHANGELOG.md
@@ -1,2 +1,25 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+
+## 0.5.0 - October 14, 2017
+
+### Added
+
+- Added support for `http-interop/http-middleware` 0.5.0
+
+### Changed
+
+- Renamed namespace to `WeCodeIn\Http\Server`
+- Changed signature of `process` method according to `http-interop/http-middleware`.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.